+++ src/steps/cloud-asset/client.ts
@@ -5,28 +5,29 @@ export class CloudAssetClient extends Client {
   private client = google.cloudasset('v1');
 
   async iterateIamPoliciesForProjectAndResources(
-    projectId: string,
     callback: (
       data: cloudasset_v1.Schema$IamPolicySearchResult,
     ) => Promise<void>,
   ): Promise<void> {
     const auth = await this.getAuthenticatedServiceClient();
 
-    await this.iterateApi(
-      async (nextPageToken) => {
-        return await this.client.v1.searchAllIamPolicies({
-          auth,
-          pageSize: 500, // 500 is the max
-          pageToken: nextPageToken,
-          scope: `projects/${projectId}`,
-        });
-      },
-      async (data: cloudasset_v1.Schema$SearchAllIamPoliciesResponse) => {
-        for (const policyResult of data.results || []) {
-          await callback(policyResult);
-        }
-      },
-    );
+    if (this.projectId) {
+      await this.iterateApi(
+        async (nextPageToken) => {
+          return await this.client.v1.searchAllIamPolicies({
+            auth,
+            pageSize: 500, // 500 is the max
+            pageToken: nextPageToken,
+            scope: `projects/${this.projectId}`,
+          });
+        },
+        async (data: cloudasset_v1.Schema$SearchAllIamPoliciesResponse) => {
+          for (const policyResult of data.results || []) {
+            await callback(policyResult);
+          }
+        },
+      );
+    }
   }
 
   // https://cloud.google.com/asset-inventory/docs/searching-iam-policies